requirement.txt
langchain openai
python-docx
python-dotenv
crewai 
crewai-tools
docx
base64

